requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
numpy
streamlit
